Add targetable unwhitelist admin command
