+++ bungee/src/main/java/com/dumbdogdiner/betterwhitelist_bungee/discord/commands/WhitelistCommand.java
@@ -2,6 +2,7 @@ package com.dumbdogdiner.betterwhitelist_bungee.discord.commands;
 
 import com.dumbdogdiner.betterwhitelist_bungee.discord.WhitelistBot;
 import com.dumbdogdiner.betterwhitelist_bungee.discord.lib.Command;
+import com.dumbdogdiner.betterwhitelist_bungee.discord.utils.RoleUtil;
 import com.dumbdogdiner.betterwhitelist_bungee.utils.PluginConfig;
 import com.dumbdogdiner.betterwhitelist_bungee.utils.SQLConnection;
 import com.dumbdogdiner.betterwhitelist_bungee.utils.UsernameValidator;
@@ -24,7 +25,7 @@ public class WhitelistCommand extends Command {
             return;
         }
 
-        if (!checkRequiredRole(e)) {
+        if (!RoleUtil.checkRequiredRole(e)) {
             e.getChannel().sendMessage(
                 ":x: **Whoops!** You don't have the role needed to be able to self-whitelist."
             ).queue();
@@ -38,7 +39,6 @@ public class WhitelistCommand extends Command {
            return;
         }
 
-        // TODO: Check why the user lookup failed.
         e.getChannel().sendTyping().queue();
 
         if (
@@ -71,79 +71,7 @@ public class WhitelistCommand extends Command {
 
         e.getChannel()
             .sendMessage(":white_check_mark: Whitelisted user **" + user.name + "**! (`" + user.id + "`)")
-            .queue();
-
-        addGrantedRole(e);
+            .queue(message -> RoleUtil.addGrantedRole(e));
     }
 
-    /**
-     * Check if a requiredRole is needed.
-     * @param e
-     * @return
-     */
-    private boolean checkRequiredRole(MessageReceivedEvent e) {
-        if (!PluginConfig.getConfig().getBoolean("discord.roles.requiredRole.enabled")) {
-            return true;
-        }
-
-        var roleId = PluginConfig.getConfig().getString("discord.roles.requiredRole.roleId");
-        var role = e.getGuild().getRoleById(roleId);
-        var member = e.getMember();
-
-        // Prevent whitelisting just in case.
-        if (role == null || member == null) {
-            WhitelistBot.getLogger().warning("[discord] Unable to retrieve requiredRole '" + roleId + "'.");
-            return false;
-        }
-
-        return member.getRoles().contains(role);
-    }
-
-    /**
-     * Add the grantedRole to users if there is one.
-     * @param e
-     */
-    private void addGrantedRole(MessageReceivedEvent e) {
-        if (PluginConfig.getConfig().getBoolean("discord.roles.grantedRole.enabled")) {
-            try {
-                var roleId = PluginConfig.getConfig().getString("discord.roles.grantedRole.roleId");
-                var role = e.getGuild().getRoleById(roleId);
-                var member = e.getMember();
-
-                // If the role doesn't exist, inform the user that something went wrong.
-                if (role == null || member == null) {
-                    WhitelistBot.getLogger().warning("[discord] Unable to retrieve grantedRole '" + roleId + "'.");
-                    sendRoleAddErrorMessage(e);
-                    return;
-                }
-
-                // Don't try and add the role if they already have it.
-                if (member.getRoles().contains(role)) {
-                    return;
-                }
-
-                // Add role to user.
-                e.getGuild().addRoleToMember(member, role).queue();
-
-                // Tell users they have been granted the role.
-                if (!PluginConfig.getConfig().getBoolean("discord.roles.grantedRole.silent")) {
-                    e.getChannel().sendMessage(":inbox_tray: You were also granted the **" + role.getName() + "** role!");
-                }
-            } catch(Exception err) {
-                err.printStackTrace();
-                sendRoleAddErrorMessage(e);
-            }
-        }
-    }
-
-    /**
-     * Send an error message telling users their granted role couldn't be added to them.
-     * @param e
-     */
-    private void sendRoleAddErrorMessage(MessageReceivedEvent e) {
-        e.getChannel()
-            .sendMessage(":warning: You are also supposed to be granted a role," +
-                "but something went wrong while I was trying to add it to you. Please contact an admin so they can add it for you!")
-            .queue();
-    }
 }
